Revert "Let assemble_vintf create the output fragments"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1dc98ebe-58ad-4297-92fe-32c2a70070f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc5d962e3f8a1ce75942327c6168929eb91f2f58