Revert "Clean up after ourselves in the Rust Benchmark Template"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e346018-a907-4585-986b-1578683fdb19). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9c9e26d08ba25c611a13938cc66d6c0117c94fb5