Redirect envsetup.sh to cogsetup.sh when in a Cog workspace
We ideally want users in Cog workspace to run cogsetup.sh, so that we can have some protections setup (e.g., not running repo command in Cog and setting up output dir symlink, setup RBE in a specific way). In order to enforce this, we need to redirect them to that script when we are in a Cog workspace.

Change-Id: Icd53c41450aff264cfcaa5516ee6c9a7e5c653ea