Revert "Ignore vndk compat symlinks in artifact path requirements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61d77720-b3ff-4009-beee-43be48675ef1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic446c260c2010a337226cf076ef753f063f2f5b5