Revert "Ignore vndk compat symlinks in artifact path requirements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02ee07a7-d3ca-4054-acca-bb73a67e7e5e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifbfe64296da2224602f4399f8168ed417b819fba