Revert "Ignore vndk compat symlinks in artifact path requirements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b1f7f7d-fd1e-4466-860a-56597cdc3867). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2aacc25164c8f9f4cfb1960d4ba8fbb143dcbe05