Revert "Ignore vndk compat symlinks in artifact path requirements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b6c164b-fc07-4300-a915-820b3ba1abd8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibee76b30c29db1626cd46ac952aafe0c0303874d