Revert "Ignore vndk compat symlinks in artifact path requirements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aa9a5629-3849-4ba5-a892-68b9a7bf1edb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia715457512695831fd712946bad8850adec60e9d