Revert "Remove TARGET_FLATTEN_APEX"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff3b6a04-3c2c-4090-b907-7c0e0373ad60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I693cdc822d50d5dbdd6188072889e47c5b1dd44c