Revert "Add 16k/4k boot OTAs to /system partition"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cfbf8e52-b5ce-4509-b000-665e73795e15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c85cb526fe6a4568ecb03b978011451ebaf5eb4