Revert "Install LOCAL_SOONG_INSTALL_SYMLINKS even if LOCAL_SOONG..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f384f3d1-848a-4afb-8ba6-6d36c83b33b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a7a65585b6d41fa917f2744d802e197907db4c5