Revert "Clean staging dirs in build_image.py"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f6ee89e-e15d-4854-9a7b-6206541e6c03). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I848282e9aa9835c9805dec09539aa1d9ca5500ef