aconfig: add exception message when fails to read from DeviceConfig
Reading value from DeviceConfig may fail if the provider is not ready.
Since DeciceConfig, and Settings class are just static wrapper on top of
the provider, it won't cause the initialization issuse. Thus when the
issue happens it means the provider is not initialized.

Test: atest aconfig.test.java and manually make a change in the
framework and check the exception message
Bug: 299471646

Change-Id: Ic08d7a9cd32a8810a7274b6d93c249abccb50b9e