Revert "Add soong config variable Release_expose_flagged_api"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e219a77c-66ba-4c4b-9652-f5edfb56a726). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I119553fdd37256de8455c0da82ad72945fc38bfb