Revert "Add soong config variable Release_expose_flagged_api"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e093df71-66ae-4eb4-baa0-754c89128ba5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I127ddd96b170a79d8a5caefaf242619fe134dcf4