Revert "Revert^2 "Add soong config variable Release_expose_flagg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c999e58-33dd-4e89-9587-bd5581f97ead). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd1bd347d9361bb33d4359c3a780b1beb0aa9b8d