Revert "Revert^2 "Add soong config variable Release_expose_flagg..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b906ac9e-e163-4549-ae55-92532184e81b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icbad966c4cd6705a1c2907786392799c90724436