Revert "Filter REL out of PLATFORM_VERSION_ALL_CODENAMES."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d67879c9-86b8-4061-b9e8-a2b412b65f2d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3e5c01ee4ab2a0828d38ee816424286c70d8233c