Revert "Add RELEASE_DEFAULT_MODULE_BUILD_FROM_SOURCE in config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7752d9ab-782f-47d3-9368-9042107c98b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0f4b36155dc33bedc4b431f97d3801844d2ac5b