Revert "Add RELEASE_DEFAULT_MODULE_BUILD_FROM_SOURCE in config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f98acf34-2f11-4003-8b3f-2a29edda1839). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I09fd56480a557c71804352376bdebafebc0d0450