Revert "Add RELEASE_DEFAULT_MODULE_BUILD_FROM_SOURCE in config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4ad03a9-46c8-44ac-8f7a-5520a7a4cd10). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4f77ad8a3e9205e9fbb3e70d6b2e889d9027457