Revert "Add RELEASE_DEFAULT_MODULE_BUILD_FROM_SOURCE in config."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af4d53ab-6d61-46b4-912a-ee32645f8dc6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f2f4277608a13bdfc493680a59b18a541c76141