release_config: build flags can be lists
Allow a build flag definition to indicate that its value should be the
concatentation of assignements, rather than the final assigned value. In
this case, the "default" value from the flag definition is always
present as the start of the list.

The initial use case for this is RELEASE_ACONFIG_VALUE_SETS, where we
need apply multiple definition files that should be processed to arrive
at the final value.

Bug: b/302593603, b/304814040
Test: manual
Change-Id: I2474cdf23341f9b1682affce6cc784281557655d