Revert "Add a check that the staging directories don't..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/19dac308-2db6-446b-8985-27301fae2940). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifa653f5ba663239ad58d37c14e75063332f66a7e