Revert "Add a check that the staging directories don't..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3cc9ed5d-9d9f-4708-84f2-544160a9fb3d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iabaf8536e085cec38e8c866deafa82ab5b64fa0b