Revert "Add a check that the staging directories don't..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a123d9e-b0a9-44d7-9e83-81e3e73285f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If38c7664f2ed3c83c64d4ad3320dd278ac1b1f27