Revert "handheld_system: Replace NFC app with NFC apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f893ea8e-a6c2-4b4d-ac06-1192a75b8424). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fa3b3c1734398f1b1e8d2fc1bfc9f501f989be2