Revert "Move most version variable out to relase configs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b5dc771-41b6-4864-93b9-d8c6bc95222c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15a1a50bc306fb3c79cb51be93520a4435f77708