Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b99dce34-9ebe-48b0-9372-012aa9ad7aa6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ided94320d24db0b7acffcea5e5d26d58a9eefd50