Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/da911fb8-0c82-45ba-a247-0f3454348752). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88beea017073a991fa9fc2623b382fea54638ad9