Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0505bc6-a056-409b-9fc0-e681f19a83bc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I59faecd2e1cf3a88f739f557ef57961f8dedf918