Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/649e0fda-5cde-49de-8e0e-4a803179deb7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I619b5f3d59fa69f54b473df257d02f4e3ffe6525