Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02d926c6-1dce-44aa-b78c-86cf29f68c1f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46b2777c32a7ba562d47c8a36fa653232b7f4e90