Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e5f6eae2-92fc-42f3-a647-d4b48c4475b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf0da1a91dc209f2a09e3e1f6bdc8c4257835120