Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d6ad210-6587-4455-a9f9-9c0c68c821cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf10dea5be1715d0691f5ede9202a5100f50b7ca