Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92e0b4f6-9f5d-43cf-ad3f-54535fcb74dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida505646e238d68fa3412458bdd29c855b7139e8