Revert "aconfig: cache flag value in generated java code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4b1ea0f-5b38-4237-a14e-838fdccc9ec8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaeacf0e83e102f26a0248318c987ea2f7012532d