Revert "Move vndk apex to the vendor partition for in-developmen..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be3a0a30-5f3d-4116-b707-5a486c9ff9a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36657ae8220bf07f3448f0520cbdd536e71b1042