Require TARGET_RELEASE for builds
This is a little complicated, in that some internal commands
intentionally set TARGET_RELEASE to an empty string (like
"lunch").  We put in logic to allow that path (and default
it to "trunk_staging").

But otherwise, we fail the build and force users to explicitly
choose what release configuration they intend for their
trunk stable build.

Bug: 307946156
Test: 'lunch' (still) works; A build attempt without `TARGET_RELEASE` set (now) fails
Change-Id: I2f667668c6688e501a3536981b7bae5fdbf962a0