Revert "Move vndk apex to the vendor partition for in-developmen..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30e34a59-29d6-4c25-b4e6-fa9e755f10da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45486bcc5b401486914f2e1e093b4f196b0805c6