Revert "Move vndk apex to the vendor partition for in-developmen..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c516f9c-821d-4c80-a238-b3c79d41d9d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b8ccb64deaf7aeb390f9d0e4a95cbc2c89bbe48