Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a433e8f9-c3db-4cf9-b88b-29dcfe16ce36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I763628bc462451585019f82c1cd9ac83725964fe