Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d277f301-c25c-4b02-95d3-da02d28047ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf19422957efa8a77a2c6e67e4bb6c311bac9da3