Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ac81df60-2600-4b1d-a88b-1dce552b822a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia063b71a3e881bdff5213c6de3c7f98a2ce9c198