Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1cf58cab-bcce-4b98-8ebf-21733480587d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17c9407fc324a862db6920e8bafba923b85960f6