Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e224689d-d6f6-497a-9726-3d94606a4672). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4320a8e42aa0e7973bd44b2df9207a9dbb0b358