Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ca6d46c-08c7-4ee6-9cf2-5b9258d8d128). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91db37186fa78d85c014172a90ca42fce2b37f6a