Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efe566ae-11f7-4e18-9fb2-454ee557fd7d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c129148b4b690de926f0214d1ff0e8ae503d2e5