Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c85ed29-03e1-4828-a33e-f013062f1024). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3eadb3abd3bf6fec90e3ee4580768a0d3fa17556