Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4617385b-e760-419d-92fe-543520f38635). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If30ee49edb7a50b54feb5ce6b4ff841c6c0a25dc