Revert "Require TARGET_RELEASE for builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e97d6df-3200-487f-8935-4e92fb44c4dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cfb57faa07c27748f40795bab86a106f7850dbc