Revert "Require TARGET_RELEASE for builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0de42c96-74e4-48e0-8aec-fb8b4b4262d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90b863d4245dfc701a262e0d0b6e32b367579fa9