Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce7ac835-9366-4eee-9561-adcb19513cf9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib3cae4f91ba55d151a39b9ae781add762742f092