Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ee46ace-fc71-4684-8cfb-4033fad31522). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16d13c519638241047c156d0244f7497e32a0976