Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1106c633-52a8-4dae-8446-89d1e9c9b7b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I553b1ca8f9c28532a1e371d808b72ed2b08d8246