Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e7407eb-ddc3-4b71-8adf-a20134742f18). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I038cc764050034176b0efc2917e8f6dccc5b6dad