Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72c152db-7a82-4b44-b419-34cf38554a1d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff5f65de2d5baed32829dcb543c109002178b8b1