Revert "Remove release_config.bzl"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/493ec569-fd4c-4552-82bc-9983bda2da1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id2c8d3f6b561ccc1adffad752433240c0073cf77