Revert "Require TARGET_RELEASE for builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/514a37bf-870e-465d-b320-790bd18debd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c5722eec0f7ad3ea57747e30cab7b5244a4772d