Rename Product variable PRODUCT_NEXT_RELEASE_HIDE_FLAGGED_API
Currently, the product variable PRODUCT_NEXT_RELEASE_HIDE_FLAGGED_API is
used to determine if the target product is a sdk variant, making the
name of the variable not too intuitive. This change fixes this by
renaming the variable to `PRODUCT_IS_SDK_VARIANT`, making the variable
more intuitive with its name and more scalable in usage.

Test: m nothing
Bug: 299570421
Change-Id: I610eae5550d940b9d33bbf37a565abcebecb5927