Revert "Define BOARD_API_LEVEL and BOARD_API_LEVEL_FROZEN"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ed6bb1e-ef62-45cc-87ae-edbdf2718e48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15ee93a02dfc698ec6f7c147a284dc7e02a5e19b