Revert "Define BOARD_API_LEVEL and BOARD_API_LEVEL_FROZEN"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83e42660-bfb1-45e8-87ef-ad969eca060b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4ac26e65b3b9a50d337adaffe66f47e3c6909635