Revert "Disable from-text stub build for sdk products"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b0bc315-9767-485f-8a19-cfd9659471b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b94f5ced7809b695fd4f9f7798dbd95212d1d5e