aconfig: add capability to create storage files part I: package.map
The package map file is a hashtable in file to store the following
mapping:

package name -> package id and flag value file start offset for this
package

So this is hashtable is used to get the starting region in the value
file that stores flag values for this package. We also store the package
id so that it is more efficient in the flags.map file which is the
second hashtable in file to be created.

1, [main.rs, commands.rs] add a new aconfig command called create-storage which takes a number
of aconfig cache files that belong to a specific container and produce
storage files.

2, [create_storage.rs] a module that executes storage creation process.
It first sort flags by package for the designated container. And then
pass the sorted flag information to each file creator.

3, [package_table.rs] package.map file creator, the PackageTabel struct
is the top level struct that represents the package hash table. The
"as_bytes" method is implemented to serialize the hash table in a fixed
format.

To ensure cross platform compatibility, we use fixed format
serialization rather in memory representation serialization. All u32 are
serialized with big endian encoding, and storage client when
deserializing would also use big endian format. Rust strings are by default UTF8
when serialized, so it is already cross platform compatible.

Note 1: unit tests for create_stroage and package_table module will be
added in next cl. Wanting to have this cl reviewed and agreed on first.
Also, try to keep cl in a reasonable size.

Note 2, flag.map and flag.val file creation will be added in further cls
using a similar pattern.

Change-Id: Ia7e9f68237ea903f295ac7891c923f6a39f3422d