Add new product property to detect if PAGE_SIZE is defined
As part of the effort to support page size agnostic builds, this patch
adds a product property that is set when Android is built with
`PRODUCT_NO_BIONIC_PAGE_SIZE_MACRO := true` set. This property is
required to be set on all devices launching with V and later.

atest -c Vts16KPageSizeTest
Bug: 315034809
Change-Id: Ib662eb82975ef36d3d3de77e9d9895f5b03c1174

Change-Id: Ia4f7cca362d8e65459b9a1f188cf60d9905c624b