aconfig: add command to export parsed_flags from cache
Introduce a new command `aconfig export-flags` to convert a set of cache
files to a `parsed_flags` protobuf message.

This used to be handled by `aconfig dump`. That command still works as
long as the `cache` and `parsed_flags` messages are identical; this CL
allows them to diverge.

Also update the build system to generate the
/<partition>/etc/aconfig_flags.pb files via `aconfig export-flags`.

Bug: 317354533
Test: atest :all
Change-Id: I462fe40d6961a3f3d9222155a750cc8949540deb