aconfig: add command to dump exported parsed_flags
This command is mainly intended for the host side implementation of
printflags, which uses the output of `m all_aconfig_declarations` as its
input. That data is generated via `aconfig export-flags`, which is a
parsed_flags protobuf message.

The new command does not support any type of --filter or --format; this
might be added in a later CL if there is sufficient need for it to
motivate the added code complexity.

Bug: 317354533
Test: printflags
Change-Id: Ie8c6c8fc29a8afaf1b62b394fb181c8bc73543d9