Revert "Don't install vintf and init.rc from Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b471ee0c-aa2c-4899-b1e1-dc0d97c562fa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I455e71702c8a025d14893ed05ea0540fe694a9db