Revert "Limit System SDK to 34 for Java modules in the vendor pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c9aca0f0-3f98-48f5-a5a0-80a323a111d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a0e66b2343552c0b841c88a4fbc0e59cbae3cd6