Revert "Limit System SDK to 34 for Java modules in the vendor pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c4b78d73-d839-449d-a1a8-6a88b170bc77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60444de1a67d90bf30858031015f92aac164c23c