Revert "Limit System SDK to 34 for Java modules in the vendor pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c2c39333-1866-4ad4-8f49-a28cca698537). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7be3bd4ddff4d30e7b705b2c7b1949c2416865d1