Revert "Limit System SDK to 34 for Java modules in the vendor pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c8ffd0f-6627-4b7f-b156-d9f8836c5534). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b3895398f5c5ead75b3689c608a2c92d8b26dc2