Revert "Limit System SDK to 34 for Java modules in the vendor pa..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f25757d-7cb7-41a0-8360-23a050ad1416). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I926da99393fe9e131ad4dfb57ec4dc523ab474a8