aconfig: create flag.map storage file
Add flag_table module for the creation of flag.map file. This is a hash
table in file similar to package.map hash table in file. This is created
per each container. It stores a mapping from a composite key (package id
and flag name) to its flag id (within package index of a flag).

This hash table allows us to look up the within package offset for a
specific flag. Along with the package offset obtained from package.map,
we will be able to find the global offset of a flag in the flag value
file.

Added unit tests to lock down the table creation and serialization.

Bug: b/312243587
Test: atest aconfig.test
Change-Id: I6e9d07bef05131cb06cfb4aa2ef29985578a6949