Add new script to create a new compat matrix for finalization
Creates the next version of the file based on the previous.
Replaces the "level" value with the new one in the file.
Adds a new Android.bp module that inherits from the previous.
Adds the current matrix file unconditional to all builds.
Adds the new matrix file contintionally based on release config.

Test: run the new script and verify the output
Bug: 279809333
Change-Id: I75075d41106fe637a3c05aaa28e21478aaad8dbf