Revert "Revert "Revert "Limit System SDK to 34 for Java modules ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9fdeabd8-d6fb-46fa-b1f2-de126b54ca6f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I12972f78f5c33c9b58e9351e58f329962e2d8c88