Revert "Revert "Revert "Limit System SDK to 34 for Java modules ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/29d8a498-cadd-44ff-b9ac-f3b3acf8566d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7c8d1ed63bfccb190eec4d404df336535e88c0b0