Fail hard if the kernel version file is used but cannot be generated.
Before this change, we generate a placeholder string for the kernel
version file to make dexppreopt happy when the kernel version file is
used but cannot be generated.

It looks like the typical case where the kernel version file cannot be
generated is when we are building GSI. Therefore, in this change, we
force `PRODUCT_ENABLE_UFFD_GC := true` in GSI so that dexpreopt doesn't
have to read the kernel version file.

After this change, we can fail hard on the missing the kernel version
file.

Test: Presubmit
Change-Id: I4207f4225560ba28861d217c0f452a56ff61a6ad