Revert "Revert "symlink_outputs: Add .KATI_SYMLINK_OUTPUTS to bu..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e92a4185-a3ac-45d3-93c3-ed1552a613ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie58a92eec48288a840de3fb634f1d4397d892657