Revert "Install libvintf_jni"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c5e526ed-25bc-43ec-a1c3-8c5be3d58cf5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb99a0dbc04a4be26d0dc4ebf5734d49dc592c93