aconfig: remove bucket_index from PackageTableNode/FlagTableNode struct
bucket index currently is a filed in PackageTableNode/FlagTableNode, but this is
purely aux info that is never searilized or deseriazlied. Therefore we
should remove it from the struct definition. Instead aconfig should
define a wrapper struct that wraps around an instance PackageTableNode/FlagTableNode
as well as aux info like bucket_index.

Bug: b/321077378
Test: atest aconfig.test; atest aconfig_storage_file.test
Change-Id: I20f2565d20b7feb5d39754e91cd6a9affb1f0e70