Revert "Rename LOCAL_USE_VNDK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bbe32cc7-d97e-4ab4-8c66-319b20c25d4b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1a38c2b094081e90a476ce93dc2c087e5f2d271