Revert "Rename LOCAL_USE_VNDK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e024402a-109b-46e5-b505-1531b25d7405). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e4ea6175af09ef0938268f30291ee70f4f78b05