Revert "Rename LOCAL_USE_VNDK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/102bb30f-9f5e-424e-bb76-e4d9cf50d4b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d68b3d8790a3dad9ddbaec4cb8d84b0c064147d