Revert "Determine GC type based on BUILT_KERNEL_VERSION_FILE."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6595a9bc-236c-4d96-aa19-816ae39d071c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd295ec24bf5d8ce9cd645350cec38f3bb4d6b5a