Revert "Use vFRC version as-is for ToT sepolicy version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9414a9f2-6745-40d6-9690-c0b8373bb4cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id18f4fcd8c7656899e10fc97e2ddaf6d22fe5028