Revert "Use vFRC version as-is for ToT sepolicy version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0bce28b4-ddda-4b52-8602-9136874aa1e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a9ce2c69f043274910e59812721bf6350729960