aconfig: introduce new flags CLI
Add a new CLI binary that we can use as the officially recommended CLI
for reading and writing flags. For now, the CLI invokes commands to
interact with DeviceConfig and aconfig protos, but we will switch it to
interact with the forthcoming mmap-file-based aconfig storage. In the
meantime, we can add improved formatting and error reporting in a
separate layer.

This CL starts with the list command, which outputs a table of the
flags.

$ adb shell flags list
namespace         package.example_flag1        A B CCCC DDDD
namespace         package.another_example_flag A B CCCC DDDD
another_namespace another_package.flag1        A B CCCC DDDD

A: T or F
B: - or O if locally overridden
C: R or W, for read-write or read-only, for system, system_ext, product,
vendor partitions
D: T or F, for default value for each partition

Bug: 324436145
Test: atest flags.test
Change-Id: Idd8aa78ccaace22a7e490014d410ad741f8f6fc6