aconfig: introduce new flags CLI
Add a new CLI binary that we can use as the officially recommended CLI
for reading and writing flags. For now, the CLI invokes commands to
interact with DeviceConfig and aconfig protos, but we will switch it to
interact with the forthcoming mmap-file-based aconfig storage. In the
meantime, we can add improved formatting and error reporting in a
separate layer.

Bug: 324436145
Test: atest flags.test
Change-Id: Idd8aa78ccaace22a7e490014d410ad741f8f6fc6