Revert "aconfig: add top level flag read lib api"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4e179ae9-a0e9-43d8-9c83-899f4a360bf7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43e41db9be099dedcb14bc77936c65f77ca37ed5