Scripts to build test-only team protos during CI
Script tools/filter_teams.py looks out output of 'all_teams' rule and
module-info to create a proto that just lists team owners for 'test'
modules (as defined by the heuristic in filter_teams).

ci/filter_teams.sh is a script to be used by the build to run filter
script.

Must be submitted with other CL in topic which removes version of script
and build rule from other repo.

Test: m all_teams filter_teams && build/make/ci/filter_teams.sh
Change-Id: I591bc5189e5075a4837c7c99fa174c6807e7a5e7