Scripts to build test-only team protos during CI
Script tools/filter_teams.py looks out output of 'all_teams' rule and
module-info to create a proto that just lists team owners for 'test'
modules (as defined by the heuristic in filter_teams).

ci/filter_teams.sh is a script to be used by the build to run filter
script.

Must be submitted with other CL in topic which removes version of script
and build rule from other repo.

After this goes in, in the past we would:
 1) remove all_teams from the current coverage build.
 2) Create a new build that builds all_teams and filter_teams.
 3) Run build/make/ci/filter_teams.sh as part of the new build after
    building the targets.

But I'm going to combine steps 2) and 3) so the "m all_teams filter_teams"
is part of the script, which is what I think was requested.
So we should:
 1) remove all_teams from the current coverage build.
 2) Create new build that does:
     rbe.set_env(environment) + 'build/make/ci/make_teams_files.sh'

Test: TARGET_PRODUCT=aosp_x86_64 TARGET_RELEASE=trunk_staging build/soong/soong_ui.bash --make-mode all_teams filter_teams && build/make/ci/make_teams_files.sh
Change-Id: I591bc5189e5075a4837c7c99fa174c6807e7a5e7