Revert "aconfig: add top level flag read lib api"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6de6d846-7fdd-48e3-b7f1-d9506a9e91b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e68d8647e94a0b72c2560b42d936da435aa1075