Revert "Do not set BOARD_VNDK_VERSION and PLATFORM_VNDK_VERSION"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61b4c01d-9c82-48f1-a1c1-89e5f6736b69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1034892d54331e73d45b3aaa33ec373fa29fd24