Revert "Fix the changing public api CUJ failure."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0f2f451-c715-45e0-9f82-e77238e5ad63). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idf45a7dbd1fe1f7a8464ca65804ab05144b4904d