aconfig: disable export mode for generated native flag  libraries
Currently the default flag values are not loaded into sysprop. Since the
current requirements for export flag are for flagged api in java and the
new storage system will load default values for all, there won't be
changes to make the default value loaded into sysprop. Thus this change
disable export mode for generated native flag libraries for now. It will
be enabled once the new storage is landed.

Test: m libaconfig_test_rust_library_with_exported_mode and expect build
failure
Bug: 327392334

Change-Id: I2db814d015cdb533778822f7417d055b251ed589