build_super_image: log LD_LIBRARY_PATH and clear it
Something invokes otatools's build_super_image with an LD_LIBRARY_PATH
containing an older copy of libc++.so, and that breaks a newer build of
lpmake, which uses a symbol from a new build of libc++.so. Clear the
variable to avoid the problem.

This is intended as a hack for libc++ testing until the real fix is
released.

Bug: none
Test: treehugger
Change-Id: I36f17abfee540ac7bc68566c01f3bd4e18b272d8