releasetools: log LD_LIBRARY_PATH and clear it
The sign_target_files_apks commands invokes the apexer, which invokes
otatools/bin/sefcontext_compile, and sefcontext_compile can't find
_ZNSt3__122__libcpp_verbose_abortEPKcz. I suspect a bad LD_LIBRARY_PATH
setting pointing to an older libc++.so.

Something else had been invoking otatools's build_super_image with an
LD_LIBRARY_PATH containing an older copy of libc++.so, and that breaks
a newer build of lpmake, which uses a symbol from a new build of
libc++.so. Clear the variable to avoid the problem. This particular
LD_LIBRARY_PATH setting seems to have been fixed now.

Bug: none
Test: treehugger
Change-Id: I36f17abfee540ac7bc68566c01f3bd4e18b272d8