Revert "Move frameworks/base/Android.mk to build/tasks"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba2a3c58-5df3-4fbf-b1c8-a05b1f319949). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic9a111fac75544ae72eb954c9c949a4f5046e501