Revert "Use soong modules for init.*.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25aee2cf-84b3-48cf-8cd4-0827dc0490f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I666bf7984c32d6e4d14a98081859076e4a833b77