Revert "Use soong modules for init.*.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25b72c41-5358-48e3-bfee-b89891cc891a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie6357cb0e79bcb150d393a28635e3cc77cbe22c9