Revert "Use soong modules for init.*.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dc4078f-b03d-4863-a37c-bf8335056a4f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dd8d2d3a3e78c93cb59ea5bab666de97b53849c