Revert "Use soong modules for init.*.rc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/186ea48b-a7c8-4a99-9ea1-4fbf2ce97063). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifc03ac40e1cd8976f09234756ee2b03612607f95