Explicitly install internal dependencies of traced_probes
traced_probes.android is a phony module that abstracts Android-only
(and optional) dependencies of traced_probes. Previously, the
dependencies were installed by directly required from traced_probes.

But that is expected to interfere with an upcoming change in Soong where
`required` dependencies are installed by Soong. Specifically, it will
increase the size of the `microdroid` module by 2-3MB because the
dependencies - which are unnecessary in Microdroid - will be installed
there due to the `required` being effective in Soong.

This change fixes that regression by making the dependencies to be only
for Android.

Bug: 321626681
Test: check the dependencies are found in Android
Test: check the dependencies are absent in Microdroid
Change-Id: I5871113e0fa6c6a9745e1c5d2e9d5c2f2dbd9764