Revert "Convert BuildIgnoreApexContritbutions variable to a boolean"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/53721efb-c49e-4ce8-b96c-fd03598b4202). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48eddc34abb188b8e24e8b2ade13af616a9a81e2